Broke out into separate module; added __reduce__ function.
